+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Resource data will be available with the resource object. This enhancement helps
   Refer example files of the updated resources for more details.
 
 #### features supported with current release
+- Connection template
 - Enclosure
 - FC network
 - Interconnect type